# By Saso Skube (34) and Tom Slejko (2)
# Via Saso Skube (3) and Skube Saso (1)
* commit '8ab53e7400f1420ca0d3ed5ed6e1e07d5ca879fc': (36 commits)
  Minor documentation/comments update
  Increased display precision of EVR prescaler readback
  Updated changelog
  Data buffer: Added windows include for endians handling
  Misc: - cleaned PSI makefile - removed not needed data buffer test file - printout fix in mrmDataBufferUser.cpp
  Added missing macros to evr vme 300 example substitution file
  SFP: bux fix (timer value was not initialized)
  Updated documentation based on recent macro changes (previous commit)
  Replaced EVR and EVG macro with DEVICE macro: - databases updates - comments updated (in scripts, substitution files, ...) - GUI now accepts macro SYS and DEVICE (separated from previous EVG/EVR macro)
  Documentation update: - Described new additions in 300 series hardware and amended to reflect recent changes in project structure - EVR manual, tutorials, readme files and comments in substitution files / startup scripts updated. - EVR and EVG startup script separated.
  Removed bunch train calculations for CMLs. Do not know where this could be used...
  Data buffer: added version check in ISR routine to make difference between hardware handling more clear
  Added missing template (delay module) to EVR build, minor readme update
  GUI: Added misc panel to EVR-PCIe (delay compensation settings)
  Data buffer: - fixed bug in remove interest, where existance of data buffer was not detected correctly - removed data buffer length define from mrfCommon, since it is not needed anywhere - added description and refractoring of test functions - misc (added comments, ...)
  Printout fixes on EVG and EVR init
  SFP: validity of readouts only printed at init and exposed to the user
  EVR event measurements: histogram record default limits increased
  GUI: - added VME-EVR-300 rear outputs - PLL lock does not exist on EVR-230 series, thus it is hidden - firmware info displayed in hex
  [COSYLAB-282] CML: added support for bunch train generation with 40 bit per event clock
  ...

+++ mrmShared/src/spi_flash.c
@@ -567,7 +567,7 @@ int spi_program_flash(void* preg, const char *bitfile)
 
 
   do {
-    size = fread(buf, 1, FLASH_DATA_ALLOC_SIZE, fd);
+    size = (int)fread(buf, 1, FLASH_DATA_ALLOC_SIZE, fd);
     flash_pageprogram(pEvr, buf, addr, size);
     if ((addr & 0x0000ffff) == 0)
       printf("\r%08x", addr);
